design is done

+++ EvacuateMe/app/src/main/java/com/example/evacuateme/Model/OrderHistory.java
@@ -43,6 +43,6 @@ public class OrderHistory {
 
     @SerializedName("rate")
     @Expose
-    public Double rate;
+    public Integer rate;
 
 }
